+++ utils.js
@@ -1,0 +1,7 @@
+export function getLocalStorage() {
+  return localStorage.getItem('lang') || 'russian';
+}
+
+export function setLocalStorage() {
+  localStorage.setItem('lang', getLocalStorage());
+}
